+++ src/system/library/multisafepay.php
@@ -1944,6 +1944,16 @@ class Multisafepay {
                 'image' => 'giropay'
             ),
             array(
+                'id' => 'GOOD4FUN',
+                'code' => 'good4fun',
+                'route' => 'multisafepay/good4fun',
+                'description' => $this->language->get('text_title_good4fun'),
+                'type' => 'giftcard',
+                'docs' => sprintf($this->language->get('text_gateway_docs_info'), ' https://docs.multisafepay.com/payment-methods/prepaid-cards/gift-cards/?utm_source=opencart&utm_medium=opencart-cms&utm_campaign=opencart-cms', $this->language->get('text_title_good4fun')),
+                'brief_description' => $this->language->get('text_brief_description_good4fun'),
+                'image' => 'good4fun'
+            ),
+            array(
                 'id' => 'GOODCARD',
                 'code' => 'goodcard',
                 'route' => 'multisafepay/goodCard',